+++ packages/browser/lib/core/crypto.ts
@@ -43,22 +43,6 @@ async function formatEncryptedData(
   )}:${base64RemovePadding(encryptedData)}:$`;
 }
 
-/**
- * Helper function to convert an offset size to a 2 byte little endian Uint8Array
- * @param number
- * @returns 2 byte little endian Uint8Array
- */
-function numberToLittleEndianUint8Array(number: number) {
-  // Create a Uint8Array with a length of 2 (assuming 16-bit integer).
-  const byteArray = new Uint8Array(2);
-
-  // Use bitwise operations to extract the bytes from the number.
-  byteArray[0] = number & 0xff; // Least significant byte
-  byteArray[1] = (number >> 8) & 0xff;
-
-  return byteArray;
-}
-
 async function formatEncryptedFileOrBlob(
   keyIv: Uint8Array,
   ecdhPublicKey: CryptoKey,
@@ -77,16 +61,14 @@ async function formatEncryptedFileOrBlob(
   ecdhPublicKey: CryptoKey,
   encryptedData: ArrayBuffer,
   isDebug: boolean,
-  fileName?: string,
-  metadata?: Uint8Array
+  fileName?: string
 ): Promise<File>;
 async function formatEncryptedFileOrBlob(
   keyIv: Uint8Array,
   ecdhPublicKey: CryptoKey,
   encryptedData: ArrayBuffer,
   isDebug: boolean,
-  fileName?: string,
-  metadata?: Uint8Array
+  fileName?: string
 ) {
   const exportableEcdhPublicKey = await window.crypto.subtle.exportKey(
     "raw",
@@ -97,12 +79,8 @@ async function formatEncryptedFileOrBlob(
   const evEncryptedFileIdentifier = new Uint8Array([
     0x25, 0x45, 0x56, 0x45, 0x4e, 0x43,
   ]);
-  const versionNumber = new Uint8Array(metadata ? [0x05] : [0x03]);
-  let offsetToData = new Uint8Array([0x37, 0x00]);
-  if (metadata) {
-    const offset = 55 + 2 + metadata.length; // 55 bytes for the header, 2 bytes for the metadata size, metadata length
-    offsetToData = numberToLittleEndianUint8Array(offset);
-  }
+  const versionNumber = new Uint8Array([0x03]);
+  const offsetToData = new Uint8Array([0x37, 0x00]);
   const flags = isDebug ? new Uint8Array([0x01]) : new Uint8Array([0x00]);
 
   const fileHeaders = concatUint8Arrays([
@@ -114,23 +92,10 @@ async function formatEncryptedFileOrBlob(
     flags,
   ]);
 
-  let fileContents: Uint8Array;
-  if (metadata) {
-    const metadataOffsetBuffer = numberToLittleEndianUint8Array(
-      metadata.length
-    );
-    fileContents = concatUint8Arrays([
-      fileHeaders,
-      metadataOffsetBuffer,
-      metadata,
-      new Uint8Array(encryptedData),
-    ]);
-  } else {
-    fileContents = concatUint8Arrays([
-      fileHeaders,
-      new Uint8Array(encryptedData),
-    ]);
-  }
+  const fileContents = concatUint8Arrays([
+    fileHeaders,
+    new Uint8Array(encryptedData),
+  ]);
 
   const crc32Hash = crc32(fileContents);
 
@@ -256,24 +221,6 @@ export class CoreCrypto {
       ["encrypt"]
     );
 
-    const metadataBytes = CoreCrypto.#buildMetadata(
-      Math.floor(Date.now() / 1000),
-      role
-    );
-    const encryptedMetadataByteBuffer = await window.crypto.subtle.encrypt(
-      {
-        name: "AES-GCM",
-        iv: keyIv,
-        tagLength: this.#config.authTagLength,
-        additionalData: this.#ecdhTeamKey,
-      },
-      derivedSecretImported,
-      metadataBytes
-    );
-    const encryptedMetadataBytes: Uint8Array = new Uint8Array(
-      encryptedMetadataByteBuffer
-    );
-
     return new Promise((resolve, reject) => {
       const reader = new window.FileReader();
 
@@ -307,8 +254,7 @@ export class CoreCrypto {
                   this.#ecdhPublicKey,
                   encrypted,
                   this.#isDebug,
-                  dataContainer.name,
-                  encryptedMetadataBytes
+                  dataContainer.name
                 )
               );
             } else {