+++ packages/ui-components/src/Form/index.tsx
@@ -1,37 +1,40 @@
-import {useEffect, useLayoutEffect, useState} from "react";
+import {useEffect, useLayoutEffect, useState } from "react";
 import { resize } from "../utilities/resize";
+import usStates from "../utilities/usStates";
 import {useMessaging} from "../utilities/useMessaging";
 import type {FormConfig, FormApiResponse, FormElement} from "./types";
 import type {EvervaultFrameHostMessages, FormFrameClientMessages} from "types";
 
 type InputRenderer = (name: string, type: string, required: boolean) => JSX.Element;
 type TextareaRenderer = (name: string, required: boolean) => JSX.Element;
-type SelectRenderer = (name: string, options: { value: string }[]) => JSX.Element;
-type FieldRenderer = InputRenderer | TextareaRenderer | SelectRenderer;
+type FieldRenderer = InputRenderer | TextareaRenderer;
 
-const fieldRenderers: Record<string, FieldRenderer> = {
-  input: (name: string, type: string, required: boolean) => (
-    <div key={name}>
-      <label htmlFor={name}>{name}</label>
-      <input type={type} name={name} id={name} required={required}/>
-    </div>
-  ),
-  textarea: (name: string, required: boolean) => (
-    <div key={name}>
-      <label htmlFor={name}>{name}</label>
-      <textarea name={name} id={name} required={required}></textarea>
-    </div>
-  ),
-  select: (name: string, options: { value: string }[]) => (
-    <div key={name}>
-      <label htmlFor={name}>{name}</label>
-      <select name={name} id={name}>
-        {options?.map((option) => (
-          <option key={option.value} value={option.value}>{option.value}</option>
+const SelectRenderer = ({ name, options }: { name: string, options: { value: string }[]}): JSX.Element => {
+  const [selectedField, setSelectedField] = useState(options[0]?.value || "");
+
+  return (
+    <div className="field-container">
+      <select name={name} value={selectedField} onChange={(e) => setSelectedField(e.target.value)}>
+        {options.map((option, i) => (
+          <option key={i} value={option.value}>{option.value}</option>
         ))}
       </select>
     </div>
+  );
+};
+
+
+const fieldRenderers: Record<string, FieldRenderer> = {
+  input: (name: string, type: string, required: boolean) => (
+    <div key={name} className="field-container">
+      <input type={type} name={name.toLocaleLowerCase()} id={name} placeholder={`${name} ${required ? "*" : "" }`} required={required}/>
+    </div>
   ),
+  textarea: (name: string, required: boolean) => (
+    <div key={name} className="field-container">
+      <textarea name={name.toLocaleLowerCase()} id={name} required={required} placeholder={`${name} ${required ? "*" : "" }`}></textarea>
+    </div>
+  )
 };
 
 export function Form({config}: { config: FormConfig }): JSX.Element {
@@ -103,8 +106,11 @@ export function Form({config}: { config: FormConfig }): JSX.Element {
           }
 
           if (element.elementType === "select") {
-            const renderField = fieldRenderers.select as SelectRenderer;
-            return renderField(element.elementName, element.options ?? []);
+            return <SelectRenderer name={element.elementName} key={element.elementName} options={element.options ?? []} />;
+          }
+
+          if (element.elementType === "select-states") {
+            return <SelectRenderer name={element.elementName} key={element.elementName} options={usStates ?? []} />;
           }
 
           return null