+++ e2e-tests/ui-components/vanilla-test-server/CHANGELOG.md
@@ -1,5 +1,18 @@
 # e2e-tests-ui-components-vanilla-server
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [408c73e]
+  - @evervault/browser@2.25.2
+
+## 0.0.14
+
+### Patch Changes
+
+- @evervault/browser@2.25.1
+
 ## 0.0.13
 
 ### Patch Changes